Move picker module from scene to framework
